requirements.txt
greenlet==3.0.1
numpy
pandas
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions
tzdata==2023.3
requests
apache-airflow
